backend/requirements.txt
fastapi
sqlmodel
sqlalchemy
psycopg2-binary
kafka-python
uvicorn
